tests/wpt/web-platform-tests/css/css-animations/parsing/animation-range-start-computed.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#propdef-animation-range">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target"></div>
<script>
test_computed_value("animation-range-start", "initial", "auto");
test_computed_value("animation-range-start", "auto");
test_computed_value("animation-range-start", "cover 0%");
test_computed_value("animation-range-start", "COVER 0%", "cover 0%");
test_computed_value("animation-range-start", "cover 100%");
test_computed_value("animation-range-start", "cover 120%");
test_computed_value("animation-range-start", "cover 42%");
test_computed_value("animation-range-start", "cover -42%");
test_computed_value("animation-range-start", "contain 42%");
test_computed_value("animation-range-start", "exit 42%");
test_computed_value("animation-range-start", "exit calc(41% + 1%)", "exit 42%");
test_computed_value("animation-range-start", "exit 1%, cover 2%, contain 100%");
</script>
